+++ packages/bpk-docs/src/routes/Routes.js
@@ -71,6 +71,7 @@ import StarRatingPage from './../pages/StarRatingPage';
 
 import AlignmentPage from './../pages/AlignmentPage';
 
+import NativeInputPage from './../pages/NativeTextInputPage';
 import NativeTextPage from './../pages/NativeTextPage';
 import NativeButtonPage from './../pages/NativeButtonPage';
 
@@ -133,6 +134,7 @@ const Routes = (
       </Route>
       <Route path={ROUTES.NATIVE_COMPONENTS}>
         <IndexRedirect to={ROUTES.NATIVE_TEXT} />
+        <Route path={ROUTES.NATIVE_INPUT} component={NativeInputPage} />
         <Route path={ROUTES.NATIVE_TEXT} component={NativeTextPage} />
         <Route path={ROUTES.NATIVE_BUTTON} component={NativeButtonPage} />
       </Route>